requirements.txt
streamlit
pandas
python-docx
PyPDF2
nano requirements.txt
